+++ src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import { createGlobalStyle } from "styled-components";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
 
-import { ScrollToTop } from "./components/ScrollToTop";
+import { ScrollToTop } from "./ScrollToTop";
 
 import { Home } from "./pages/Home";
 import { ExperienceDetail } from "./pages/ExperienceDetail";
